+++ AutoStack/Patches_Chests.cs
@@ -1,9 +1,4 @@
-﻿using Shared;
-using System.Reflection;
-using System.Reflection.Emit;
-using UnityEngine;
-
-namespace AutoStackNS
+﻿namespace AutoStackNS
 {
     /// <summary>
     /// Makes ResourceChest always count as 1 card (like when playing on the cities board).
@@ -28,6 +23,25 @@ namespace AutoStackNS
             {
                 return true;
             }
+        }
+    }
+
+    [HarmonyPatch(typeof(ResourceChest), nameof(ResourceChest.CanHaveCard))]
+    public static class Patch_ResourceChestAllowed
+    {
+        public static bool Prefix(CardData otherCard, ResourceChest __instance, ref bool __result)
+        {
+            if (__instance.HeldCardId is not (null or "") && otherCard.Id != __instance.HeldCardId)
+                __result = false;
+            else if (otherCard.Id == "gold" || otherCard.Id == "shell" || otherCard is Dollar)
+                __result = false;
+            else if (otherCard is Food or Equipable)
+                __result = true;
+            else if (otherCard.MyCardType is CardType.Resources)
+                __result = true;
+            else
+                __result = false;
+            return false;
         }
     }
 }