+++ AutoStack/changelog.md
@@ -1,4 +1,12 @@
 # Changelog
 
+## [1.0.1]
+- added: shift dragging resource chests will dispense and move its contents instead
+- added: Hotpots act like chests for all food (item magnet)
+- added: allow Hotpot to work while parent is Mess Hall
+- added: animals stay in breeding pen
+- added: Resource Chest allow more (food, equipment)
+- added: press shift while equipping items to keep villager type (mage, archer, ...)
+
 ## [1.0.0]
 - first release